Split API generation helpers
